+++ CHANGELOG.md
@@ -6,6 +6,9 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Added
+- Added a JSON representation of the response payload when requesting fox sports matchup data
+- Created a data model that represents fox sports matchup
 
 ## [0.3.0] - 2025-04-04
 ### Changed